优化购物车

+++ A3Mall/mall/basic/Shopping.php
@@ -12,7 +12,7 @@ use mall\utils\BC;
 use mall\utils\Tool;
 use think\facade\Db;
 
-class Cart {
+class Shopping {
 
     public static function get($cart){
         $data = [
@@ -100,7 +100,77 @@ class Cart {
         return $data;
     }
 
+    public static function add($id,$sku_id,$num=1){
+        $num = $num ?: 1;
+
+        $goods = Db::name("goods")->where("id",$id)->find();
+        if(empty($goods) || $goods["status"] != 0){
+            throw new \Exception("该商品己下架！",0);
+        }
+
+        $products = null;
+        if(Db::name("goods_item")->where("goods_id",$id)->count()){
+            $products = Db::name("goods_item")->where([
+                "id"=>$sku_id,
+                "goods_id"=>$id
+            ])->find();
+
+            if(empty($products)){
+                throw new \Exception("该商品己下架。",0);
+            }
+        }
+
+        if(!empty($products)){
+            if($products["store_nums"] < $num){
+                throw new \Exception("您选择的商品库存不足",0);
+            }
+        }else{
+            if($goods["store_nums"] < $num){
+                throw new \Exception("您选择的商品库存不足",0);
+            }
+        }
+
+        $cart = [
+            "session_id"=>session_id(),
+            "user_id"=>Users::get("id"),
+            "goods_id"=>$id,
+            "product_id"=>!empty($products["id"]) ? $products["id"] : 0,
+            "spec_key"=>!empty($products["spec_key"]) ? $products["spec_key"] : "",
+            "sell_price"=>!empty($products["sell_price"]) ? $products["sell_price"] : $goods["sell_price"],
+            "cost_price"=>!empty($products["cost_price"]) ? $products["cost_price"] : $goods["cost_price"],
+            "market_price"=>!empty($products["market_price"]) ? $products["market_price"] : $goods["market_price"],
+            "goods_weight"=>!empty($products["goods_weight"]) ? $products["goods_weight"] : $goods["goods_weight"],
+            "goods_nums"=>$num,
+            "create_time"=>time()
+        ];
+
+        $map = [
+            "goods_id"=>$id,
+            "user_id"=>Users::get("id")
+        ];
+
+        if(!empty($products["spec_key"])){
+            $map["spec_key"] = $products["spec_key"];
+        }
+
+        if(Db::name("cart")->where($map)->count()){
+            Db::name("cart")->where($map)->update([
+                "goods_nums"=>$num,
+                "update_time"=>time()
+            ]);
+        }else{
+            Db::name("cart")->insert($cart);
+        }
+
+        return true;
+    }
+
     public static function delete($id){
+        if(empty($id)){
+            throw new \Exception("非法参数",0);
+        }
+
+        $id = array_map("intval",explode(",",$id));
         return Db::name("cart")->where("id","in",$id)->delete();
     }
 }
